+++ kernel/raster/psf.h
@@ -1,12 +1,12 @@
 #pragma once
-#include <stdint.h>
 #include <stddef.h>
+#include <stdint.h>
 
 extern uint8_t _binary_kernelfont_psf_start[];
 extern uint8_t _binary_kernelfont_psf_end[];
 
 void psf_init(void);
-size_t psf_getwidth(void);
-size_t psf_getheight(void);
+int psf_getwidth(void);
+int psf_getheight(void);
 size_t psf_getbytesperline(void);
 uint8_t *psf_getglyph(uint32_t chr);
